refactoring: change type of 'amount' for m68k bus

+++ m68k_emu/src/bus.rs
@@ -1,4 +1,4 @@
 pub trait BusM68k {
-    fn read(&self, address: u32, amount: u32) -> Result<u32, ()>;
-    fn write(&mut self, data: u32, address: u32, amount: u32) -> Result<(), ()>;
+    fn read(&self, address: u32, amount: usize) -> Result<u32, ()>;
+    fn write(&mut self, data: u32, address: u32, amount: usize) -> Result<(), ()>;
 }
